Reorganize and cleanup truffle integration tests
