+++ .circleci/config.yml
@@ -78,7 +78,8 @@ workflows:
   build:
     jobs:
       - unit-test
-      - e2e-zeppelin
+      # TODO: re-enable when server logic is added
+      #- e2e-zeppelin
       - e2e-metacoin
   # TODO: re-enable.
   # At the moment we're using forks so this is pointless
